init.lua: Add gf mapping for tar files
Open files from tarballs by saving them to a temporary file with the
same extension and then opening that. This is what netrw does with
remote files. The tar.vim plugin doesn't do that. It tries to :read the
files from the output of the tar command. This fails miserably for
binary files. That strategy can be made to work, but the better strategy
that works in all cases is saving to a temporary file and then opening
that. That allows all normal file handling to proceed.